Arreglado coeficientes sin -a
Durante la creación cambie MinAbs sin pensarlo y rompió los resultados de las reglas de coeficientes óptimas
